\name{distValues}
\alias{distValues}
\docType{data}
\title{
distValues
}
\description{
Returns a data.frame of distribution-level test values from a CuffDist object (@promoters, @splicing, @relCDS)
}
\usage{data(distValues)}
\format{
}
\details{
None
}
\source{

}
\references{

}
\examples{

}
\keyword{datasets}
