\name{sampleGeneSet}
\alias{sampleGeneSet}
\docType{data}
\title{
sampleGeneSet
}
\description{
A sample CuffGeneSet data set for 20 genes.
}
\usage{data(sampleData)}
\format{
  The format is:
Formal class 'CuffGeneSet' [package "cummeRbund"] with 7 slots
  ..@ ids       : chr "c(\"XLOC_000006\", \"XLOC_000007\", \"XLOC_000035\", \"XLOC_000067\", \"XLOC_000068\", \"XLOC_000124\", \"XLOC_000130\", \"XLOC"| __truncated__
  ..@ isoforms  :Formal class 'CuffFeatureSet' [package "cummeRbund"] with 3 slots
  .. .. ..@ annotation:'data.frame':	53 obs. of  16 variables:
  .. .. .. ..$ isoform_id     : chr [1:53] "TCONS_00000010" "TCONS_00000011" "TCONS_00000062" "TCONS_00000063" ...
  .. .. .. ..$ gene_id        : chr [1:53] "XLOC_000006" "XLOC_000007" "XLOC_000035" "XLOC_000035" ...
  .. .. .. ..$ CDS_id         : chr [1:53] "NA" "NA" "NA" "NA" ...
  .. .. .. ..$ TSS_group_id   : chr [1:53] NA "TSS4" "TSS31" "TSS32" ...
  .. .. .. ..$ class_code     : chr [1:53] "=" "=" "=" "=" ...
  .. .. .. ..$ nearest_ref_id : chr [1:53] "uc010nxu.1" "uc001aax.1" "uc001afv.2" "uc001afw.2" ...
  .. .. .. ..$ locus          : chr [1:53] "chr1:367658-368595" "chr1:420205-421839" "chr1:1407163-1431581" "chr1:1407163-1431581" ...
  .. .. .. ..$ length         : int [1:53] 937 802 2432 1939 2667 1170 1248 687 1863 2015 ...
  .. .. .. ..$ coverage       : num [1:53] NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..$ gene_id        : chr [1:53] "XLOC_000006" "XLOC_000007" "XLOC_000035" "XLOC_000035" ...
  .. .. .. ..$ class_code     : chr [1:53] NA NA NA NA ...
  .. .. .. ..$ nearest_ref_id : chr [1:53] NA NA NA NA ...
  .. .. .. ..$ gene_short_name: chr [1:53] "OR4F16" NA "ATAD3B" "ATAD3B" ...
  .. .. .. ..$ locus          : chr [1:53] "chr1:367658-368595" "chr1:420205-421839" "chr1:1407163-1431581" "chr1:1407163-1431581" ...
  .. .. .. ..$ length         : int [1:53] NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..$ coverage       : num [1:53] NA NA NA NA NA NA NA NA NA NA ...
  .. .. ..@ fpkm      :'data.frame':	159 obs. of  6 variables:
  .. .. .. ..$ isoform_id  : chr [1:159] "TCONS_00000010" "TCONS_00000011" "TCONS_00000062" "TCONS_00000063" ...
  .. .. .. ..$ sample_name : Factor w/ 3 levels "hESC","Fibroblasts",..: 2 2 2 2 2 2 2 2 2 2 ...
  .. .. .. ..$ fpkm        : num [1:159] 0 0 195.6 49.7 135.2 ...
  .. .. .. ..$ conf_hi     : num [1:159] 0 0 306 102 222 ...
  .. .. .. ..$ conf_lo     : num [1:159] 0 0 85 0 48 ...
  .. .. .. ..$ quant_status: chr [1:159] "OK" "OK" "OK" "OK" ...
  .. .. ..@ diff      :'data.frame':	159 obs. of  11 variables:
  .. .. .. ..$ isoform_id    : chr [1:159] "TCONS_00000010" "TCONS_00000011" "TCONS_00000062" "TCONS_00000063" ...
  .. .. .. ..$ sample_1      : Factor w/ 3 levels "hESC","Fibroblasts",..: 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ sample_2      : Factor w/ 3 levels "hESC","Fibroblasts",..: 2 2 2 2 2 2 2 2 2 2 ...
  .. .. .. ..$ status        : chr [1:159] "NOTEST" "NOTEST" "OK" "OK" ...
  .. .. .. ..$ value_1       : num [1:159] 0 0 669 191 217 ...
  .. .. .. ..$ value_2       : num [1:159] 0 0 195.6 49.7 135.2 ...
  .. .. .. ..$ ln_fold_change: num [1:159] 0 0 -1.773 -1.939 -0.684 ...
  .. .. .. ..$ test_stat     : num [1:159] 0 0 4.18 2.51 1.39 ...
  .. .. .. ..$ p_value       : num [1:159] 1.00 1.00 2.91e-05 1.21e-02 1.65e-01 ...
  .. .. .. ..$ q_value       : num [1:159] 1 1 0.000286 0.048893 0.363223 ...
  .. .. .. ..$ significant   : chr [1:159] "no" "no" "yes" "yes" ...
  ..@ TSS       :Formal class 'CuffFeatureSet' [package "cummeRbund"] with 3 slots
  .. .. ..@ annotation:'data.frame':	27 obs. of  14 variables:
  .. .. .. ..$ TSS_group_id   : chr [1:27] "TSS173" "TSS189" "TSS190" "TSS1930" ...
  .. .. .. ..$ class_code     : chr [1:27] NA NA NA NA ...
  .. .. .. ..$ nearest_ref_id : chr [1:27] NA NA NA NA ...
  .. .. .. ..$ gene_id        : chr [1:27] "XLOC_000124" "XLOC_000130" "XLOC_000130" "XLOC_001235" ...
  .. .. .. ..$ locus          : chr [1:27] "chr1:13641972-13648988" "chr1:15479027-15546973" "chr1:15479027-15546973" "chr1:1108435-1133313" ...
  .. .. .. ..$ length         : int [1:27] NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..$ coverage       : num [1:27] NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..$ gene_id        : chr [1:27] "XLOC_000124" "XLOC_000130" "XLOC_000130" "XLOC_001235" ...
  .. .. .. ..$ class_code     : chr [1:27] NA NA NA NA ...
  .. .. .. ..$ nearest_ref_id : chr [1:27] NA NA NA NA ...
  .. .. .. ..$ gene_short_name: chr [1:27] "PRAMEF9" "TMEM51" "TMEM51" NA ...
  .. .. .. ..$ locus          : chr [1:27] "chr1:13641972-13648988" "chr1:15479027-15546973" "chr1:15479027-15546973" "chr1:1108435-1133313" ...
  .. .. .. ..$ length         : int [1:27] NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..$ coverage       : num [1:27] NA NA NA NA NA NA NA NA NA NA ...
  .. .. ..@ fpkm      :'data.frame':	81 obs. of  6 variables:
  .. .. .. ..$ TSS_group_id: chr [1:81] "TSS173" "TSS189" "TSS190" "TSS1930" ...
  .. .. .. ..$ sample_name : Factor w/ 3 levels "hESC","Fibroblasts",..: 2 2 2 2 2 2 2 2 2 2 ...
  .. .. .. ..$ fpkm        : num [1:81] 0 26.43 681.57 0 2.84 ...
  .. .. .. ..$ conf_hi     : num [1:81] 0 315.5 974.6 0 5.7 ...
  .. .. .. ..$ conf_lo     : num [1:81] 0 0 389 0 0 ...
  .. .. .. ..$ quant_status: chr [1:81] "OK" "OK" "OK" "OK" ...
  .. .. ..@ diff      :'data.frame':	81 obs. of  11 variables:
  .. .. .. ..$ TSS_group_id  : chr [1:81] "TSS173" "TSS189" "TSS190" "TSS1930" ...
  .. .. .. ..$ sample_1      : Factor w/ 3 levels "hESC","Fibroblasts",..: 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ sample_2      : Factor w/ 3 levels "hESC","Fibroblasts",..: 2 2 2 2 2 2 2 2 2 2 ...
  .. .. .. ..$ status        : chr [1:81] "NOTEST" "OK" "OK" "OK" ...
  .. .. .. ..$ value_1       : num [1:81] 0 114 269.1 9.7 19.4 ...
  .. .. .. ..$ value_2       : num [1:81] 0 26.43 681.57 0 2.84 ...
  .. .. .. ..$ ln_fold_change: num [1:81] 0.00 -2.11 1.34 -1.80e+308 -2.77 ...
  .. .. .. ..$ test_stat     : num [1:81] 0.00 2.67e-01 -3.17 -1.80e+308 3.62 ...
  .. .. .. ..$ p_value       : num [1:81] 1.00 7.90e-01 1.51e-03 2.79e-06 2.96e-04 ...
  .. .. .. ..$ q_value       : num [1:81] 1.00 9.39e-01 5.87e-03 2.13e-05 1.41e-03 ...
  .. .. .. ..$ significant   : chr [1:81] "no" "no" "yes" "yes" ...
  ..@ CDS       :Formal class 'CuffFeatureSet' [package "cummeRbund"] with 3 slots
  .. .. ..@ annotation:'data.frame':	23 obs. of  15 variables:
  .. .. .. ..$ CDS_id         : chr [1:23] "P129" "P252" "P270" "P271" ...
  .. .. .. ..$ class_code     : chr [1:23] NA NA NA NA ...
  .. .. .. ..$ nearest_ref_id : chr [1:23] NA NA NA NA ...
  .. .. .. ..$ gene_id        : chr [1:23] "XLOC_000068" "XLOC_000124" "XLOC_000130" "XLOC_000130" ...
  .. .. .. ..$ TSS_group_id   : chr [1:23] "TSS92" "TSS173" "TSS190" "TSS189,TSS190" ...
  .. .. .. ..$ locus          : chr [1:23] "chr1:6304261-6305638" "chr1:13641972-13648988" "chr1:15479027-15546973" "chr1:15479027-15546973" ...
  .. .. .. ..$ length         : int [1:23] NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..$ coverage       : num [1:23] NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..$ gene_id        : chr [1:23] "XLOC_000068" "XLOC_000124" "XLOC_000130" "XLOC_000130" ...
  .. .. .. ..$ class_code     : chr [1:23] NA NA NA NA ...
  .. .. .. ..$ nearest_ref_id : chr [1:23] NA NA NA NA ...
  .. .. .. ..$ gene_short_name: chr [1:23] "HES3" "PRAMEF9" "TMEM51" "TMEM51" ...
  .. .. .. ..$ locus          : chr [1:23] "chr1:6304261-6305638" "chr1:13641972-13648988" "chr1:15479027-15546973" "chr1:15479027-15546973" ...
  .. .. .. ..$ length         : int [1:23] NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..$ coverage       : num [1:23] NA NA NA NA NA NA NA NA NA NA ...
  .. .. ..@ fpkm      :'data.frame':	69 obs. of  6 variables:
  .. .. .. ..$ CDS_id      : chr [1:69] "P129" "P252" "P270" "P271" ...
  .. .. .. ..$ sample_name : Factor w/ 3 levels "hESC","Fibroblasts",..: 2 2 2 2 2 2 2 2 2 2 ...
  .. .. .. ..$ fpkm        : num [1:69] 0 0 421.374 286.625 0.625 ...
  .. .. .. ..$ conf_hi     : num [1:69] 0 0 651.53 444.92 3.37 ...
  .. .. .. ..$ conf_lo     : num [1:69] 0 0 191 128 0 ...
  .. .. .. ..$ quant_status: chr [1:69] "OK" "OK" "OK" "OK" ...
  .. .. ..@ diff      :'data.frame':	69 obs. of  11 variables:
  .. .. .. ..$ CDS_id        : chr [1:69] "P129" "P252" "P270" "P271" ...
  .. .. .. ..$ sample_1      : Factor w/ 3 levels "hESC","Fibroblasts",..: 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ sample_2      : Factor w/ 3 levels "hESC","Fibroblasts",..: 2 2 2 2 2 2 2 2 2 2 ...
  .. .. .. ..$ status        : chr [1:69] "OK" "NOTEST" "OK" "OK" ...
  .. .. .. ..$ value_1       : num [1:69] 2328.4 0 108.4 274.7 2.6 ...
  .. .. .. ..$ value_2       : num [1:69] 0 0 421.374 286.625 0.625 ...
  .. .. .. ..$ ln_fold_change: num [1:69] -1.80e+308 0.00 1.96 6.14e-02 -2.05 ...
  .. .. .. ..$ test_stat     : num [1:69] -1.80e+308 0.00 -4.55 -1.38e-01 5.77e-01 ...
  .. .. .. ..$ p_value       : num [1:69] 1.41e-33 1.00 5.49e-06 8.90e-01 5.64e-01 ...
  .. .. .. ..$ q_value       : num [1:69] 9.72e-32 1.00 4.48e-05 1.06 7.52e-01 ...
  .. .. .. ..$ significant   : chr [1:69] "yes" "no" "yes" "no" ...
  ..@ annotation:'data.frame':	20 obs. of  7 variables:
  .. ..$ gene_id        : chr [1:20] "XLOC_000006" "XLOC_000007" "XLOC_000035" "XLOC_000067" ...
  .. ..$ class_code     : chr [1:20] NA NA NA NA ...
  .. ..$ nearest_ref_id : chr [1:20] NA NA NA NA ...
  .. ..$ gene_short_name: chr [1:20] "OR4F16" NA "ATAD3B" "C1orf211" ...
  .. ..$ locus          : chr [1:20] "chr1:367658-368595" "chr1:420205-421839" "chr1:1407163-1431581" "chr1:6297870-6299490" ...
  .. ..$ length         : int [1:20] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ coverage       : num [1:20] NA NA NA NA NA NA NA NA NA NA ...
  ..@ fpkm      :'data.frame':	60 obs. of  6 variables:
  .. ..$ gene_id     : chr [1:60] "XLOC_000006" "XLOC_000007" "XLOC_000035" "XLOC_000067" ...
  .. ..$ sample_name : Factor w/ 3 levels "hESC","Fibroblasts",..: 2 2 2 2 2 2 2 2 2 2 ...
  .. ..$ fpkm        : num [1:60] 0 0 650.64 4.36 0 ...
  .. ..$ conf_hi     : num [1:60] 0 0 880.01 9.93 0 ...
  .. ..$ conf_lo     : num [1:60] 0 0 421 0 0 ...
  .. ..$ quant_status: chr [1:60] "OK" "OK" "OK" "OK" ...
  ..@ diff      :'data.frame':	60 obs. of  11 variables:
  .. ..$ gene_id       : chr [1:60] "XLOC_000006" "XLOC_000007" "XLOC_000035" "XLOC_000067" ...
  .. ..$ sample_1      : Factor w/ 3 levels "hESC","Fibroblasts",..: 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ sample_2      : Factor w/ 3 levels "hESC","Fibroblasts",..: 2 2 2 2 2 2 2 2 2 2 ...
  .. ..$ status        : chr [1:60] "NOTEST" "NOTEST" "OK" "OK" ...
  .. ..$ value_1       : num [1:60] 0 0 1598 216 2328 ...
  .. ..$ value_2       : num [1:60] 0 0 650.64 4.36 0 ...
  .. ..$ ln_fold_change: num [1:60] 0.00 0.00 -1.30 -5.63 -1.80e+308 ...
  .. ..$ test_stat     : num [1:60] 0.00 0.00 4.82 6.00 -1.80e+308 ...
  .. ..$ p_value       : num [1:60] 1.00 1.00 1.43e-06 1.99e-09 1.41e-33 ...
  .. ..$ q_value       : num [1:60] 1.00 1.00 1.05e-05 2.17e-08 7.57e-32 ...
  .. ..$ significant   : chr [1:60] "no" "no" "yes" "yes" ...
}
\details{
None
}
\source{
None
}
\references{
None
}
\examples{
data(sampleData)

}
\keyword{datasets}
